Update isotope service (#12)
* Capture errors in readAllAndClose

* Respond with byte array rather than echo

* Return an error if response from child service is not 200

* Remove duplicate error logging

* Remove naked return in executeRequestCommand

* Add missing return

* Make a random payload used across responses

* Make a random payload for each request

+++ isotope/service/pkg/srv/prometheus/handler.go
@@ -96,7 +96,7 @@ func RecordRequestSent(destinationService string, size uint64) {
 
 // RecordResponseSent observes the time-to-response duration and size for the
 // HTTP status code.
-func RecordResponseSent(duration time.Duration, size uint64, code int) {
+func RecordResponseSent(duration time.Duration, size int, code int) {
 	strCode := strconv.Itoa(code)
 	serviceRequestDurationSeconds.WithLabelValues(strCode).Observe(
 		duration.Seconds())
